multihash/requirements.txt
pycrypto
bitarray